copy over schema
