+++ code/cell.db/package.json
@@ -31,5 +31,6 @@
     "!*.TEST.{js, mjs, d.ts}",
     "!*.INTEGRATION.{js, mjs, d.ts}"
   ],
+  "repository": "uiharness/platform",
   "license": "MIT"
 }
