+++ code/cell.func.sys/package.json
@@ -28,5 +28,6 @@
     "!*.TEST.{js, mjs, d.ts}",
     "!*.INTEGRATION.{js, mjs, d.ts}"
   ],
+  "repository": "uiharness/platform",
   "license": "MIT"
 }
