+++ code/cell.http/package.json
@@ -32,5 +32,6 @@
     "!*.TEST.{js, mjs, d.ts}",
     "!*.INTEGRATION.{js, mjs, d.ts}"
   ],
+  "repository": "uiharness/platform",
   "license": "MIT"
 }
