+++ code/cell.schema/package.json
@@ -12,7 +12,11 @@
     "prepare": "ts prepare"
   },
   "dependencies": {
-    "@platform/libs": "0.1.3"
+    "@platform/cell.coord": "0.5.29",
+    "@platform/cell.types": "0.0.26",
+    "@platform/fsdb.types": "0.3.10",
+    "@platform/libs": "0.1.3",
+    "@platform/util.value": "0.5.5"
   },
   "devDependencies": {
     "@platform/test": "0.0.50",
@@ -24,5 +28,6 @@
     "!*.TEST.{js, mjs, d.ts}",
     "!*.INTEGRATION.{js, mjs, d.ts}"
   ],
+  "repository": "uiharness/platform",
   "license": "MIT"
 }
