+++ code/cell.schema/src/common/libs.cell.ts
@@ -1,0 +1,2 @@
+export { coord } from '@platform/cell.coord';
+// export { value } from '@platform/cell.value';
