+++ code/cell.schema/src/common/libs.ts
@@ -1,2 +1,7 @@
 import { uniq } from 'ramda';
 export const R = { uniq };
+
+export { coord } from './libs.cell';
+
+import { id } from '@platform/util.value';
+export const cuid = id.cuid;
