+++ code/cell.schema/src/common/types.ts
@@ -1,1 +1,2 @@
-export * from '../types';
+export * from '@platform/cell.types';
+export * from '@platform/fsdb.types';
