+++ code/cell.schema/src/test/index.ts
@@ -1,0 +1,3 @@
+export { expect, expectError } from '@platform/test';
+export { time } from '@platform/util.value';
+export * from '../common';
