+++ code/cell.schema/src/types.ts
@@ -1,1 +1,1 @@
-export type TEMP = string;
+export * from '@platform/cell.types/lib/types/types.uri';
