+++ code/cell.types/package.json
@@ -15,6 +15,7 @@
   },
   "dependencies": {
     "@platform/cache": "0.2.0",
+    "@platform/fsdb.types": "0.3.10",
     "@platform/libs": "0.1.3",
     "@platform/util.diff": "0.0.1"
   },
@@ -28,5 +29,6 @@
     "!*.TEST.{js, mjs, d.ts}",
     "!*.INTEGRATION.{js, mjs, d.ts}"
   ],
+  "repository": "uiharness/platform",
   "license": "MIT"
 }
