+++ code/cell.types/src/types/index.ts
@@ -1,6 +1,7 @@
 export * from './types.cell';
 export * from './types.cell.coord';
 export * from './types.context';
+export * from './types.db';
 export * from './types.error';
 export * from './types.func';
 export * from './types.get';
