+++ code/fsdb.model/package.json
@@ -27,5 +27,6 @@
   "files": [
     "lib"
   ],
+  "repository": "uiharness/platform",
   "license": "MIT"
 }
